refactor(render): wire component library -- kill shadows, promote context components, migrate separators
Three improvements to component library utilization:

1. Kill local statusBadge shadow in task-view-mode.ts -- replace with
   canonical import from render/status-badge.ts + taskStatusToBadge mapper

2. Promote context-bar.ts and context-sparkline.ts from cli/ root into
   render/ component library and re-export from barrel index

3. Add optional maxWidth parameter to divider() and migrate 8 inline
   separator sites to use the component instead of hand-rolling
   palette.dim + repeat:
   - task-view.ts (2 sites, cap 100)
   - turn-record-renderer.ts (2 sites, cap 100)
   - task-view-mode.ts (2 sites, cap 80)
   - plugin-agents.ts (1 site, cap 60)
   - changelog.ts (2 sites, cap 60)
