+++ src/cli/render/context-sparkline.test.ts
