+++ src/cli/render/context-sparkline.ts
